feat: context-aware channel rate limit with float RPM support
- Context-aware skip (other channels available) or 429 (all rate-limited)
- RateLimitRPM/DefaultRPM changed from int to float64, window auto-scales
- controller/relay.go: skip rate-limited channels via ExcludeChannel
- middleware/distributor.go: remove reject, defer to controller
- constant/context_key.go: add rate limit tracking keys
- relaykit/types/error.go: add ErrorCodeChannelRateLimited
- model/channel_cache.go: add GetChannelIDsForGroupModel

+++ web/src/features/channels/lib/channel-form.ts
@@ -281,7 +281,7 @@ export const channelFormSchema = z
     claude_beta_query: z.boolean().optional(), // Anthropic: beta query passthrough
     disable_task_polling_sleep: z.boolean().optional(),
     rate_limit_enabled: z.boolean().optional(),
-    rate_limit_rpm: z.number().int().min(0).optional(),
+    rate_limit_rpm: z.number().min(0).optional(),
     rate_limit_tpm: z.number().int().min(0).optional(),
     // Upstream model update settings (stored in settings JSON)
     upstream_model_update_check_enabled: z.boolean().optional(),
